third_party/WebKit/LayoutTests/platform/mac-mac10.11/external/wpt/dom/events/Event-timestamp-high-resolution-expected.txt

This is a testharness.js-based test.
FAIL Constructed MouseEvent timestamp should be high resolution and have the same time origin as performance.now() assert_greater_than_equal: Event timestamp should be greater than performance.now() timestamp taken before its creation expected a number greater than or equal to 19.20000000973232 but got 19.199999995180406
PASS Constructed KeyboardEvent timestamp should be high resolution and have the same time origin as performance.now()
PASS Constructed WheelEvent timestamp should be high resolution and have the same time origin as performance.now()
PASS Constructed GamepadEvent timestamp should be high resolution and have the same time origin as performance.now()
PASS Constructed FocusEvent timestamp should be high resolution and have the same time origin as performance.now()
Harness: the test ran to completion.
